Commit 1
Made the key by combining all the ASCII values, to handle collisions I just made it so that it moved on to the next index. Chose that method because it seemed like it was the simplest, so figured I'd get it out of the way before moving onto more complicated methods. Results for the title table was 13328 collisions with no empty space, for the quote table it was 13770 collisions with no empty space, and the creating both tables took 53.72 seconds.