Commit 4
I changed back to just filling the list, and not using a linked list. I also changed my original hash function, and added a second to go with the "double hash" method. This method was actually the worst so far, with both tables having 14,995 collisions, and the title table having 1 empty space and quote table having no empty spaces, and it took over 200 seconds to complete.