Commit 5
For this one, I went back to using the djbHash, and I made each element of the array into a linked list, but made it so that they had a max length of 2, and if it had 2 elements in the linked list, it would move on and test that one. This method had 9687 collisions in the title table with 286 empty spaces, 8929 collisions on the quote table with 231 empty spaces, and both tables took 0.85 seconds to construct.